+++ src/misc/providers.ts
@@ -1,4 +1,4 @@
-import {EnvironmentProviders, Provider, makeEnvironmentProviders} from '@angular/core';
+import {EnvironmentProviders, InjectionToken, Provider, Type, makeEnvironmentProviders} from '@angular/core';
 import {isPresent} from '@jscrpt/common';
 
 import {GLOBAL_NOTIFICATIONS, LOCAL_NOTIFICATIONS} from '../common/notifications.service';
@@ -6,14 +6,15 @@ import {GLOBAL_NOTIFICATIONS, LOCAL_NOTIFICATIONS} from '../common/notifications
 /**
  * Provides global notifications service
  * @param name - Name for global notifications scope
+ * @param customNotificationsToken - Optional type or token that should be provided for custom notifications
  */
-export function provideGlobalNotifications(name?: string): EnvironmentProviders
+export function provideGlobalNotifications(name?: string, customNotificationsToken?: Function | Type<any> | InjectionToken<unknown>): EnvironmentProviders
 {
     if(isPresent(name))
     {
         return makeEnvironmentProviders(
         [
-            GLOBAL_NOTIFICATIONS.named(name),
+            GLOBAL_NOTIFICATIONS.named(name, customNotificationsToken),
         ]);
     }
 
@@ -26,13 +27,14 @@ export function provideGlobalNotifications(name?: string): EnvironmentProviders
 /**
  * Provides local notifications service
  * @param name - Name for local notifications scope
+ * @param customNotificationsToken - Optional type or token that should be provided for custom notifications
  */
-export function provideLocalNotifications(name?: string): Provider[]
+export function provideLocalNotifications(name?: string, customNotificationsToken?: Function | Type<any> | InjectionToken<unknown>): Provider[]
 {
     if(isPresent(name))
     {
         return [
-            LOCAL_NOTIFICATIONS.named(name),
+            LOCAL_NOTIFICATIONS.named(name, customNotificationsToken),
         ];
     }